remove getMetrics function and add to own file
